pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "flux-titan"
version = "0.2.0"
description = "Self-hosted AI newsroom for Telegram channels with AI summarization"
authors = [
    { name = "Flux-Titan Contributors" }
]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "feedparser>=6.0",
    "google-generativeai>=0.5",
    "python-telegram-bot>=21.0",
    "python-dotenv>=1.0",
    "aiohttp>=3.9",
    "httpx>=0.25",
    "openai>=1.0",
    "PyYAML>=6.0",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0"
]

[project.scripts]
flux-titan = "flux_titan.cli:run_cli"
